+++ upcloud/kubernetes_test.go
@@ -14,6 +14,7 @@ const exampleKubernetesClusterJSON string = `{
 	"state": "running",
 	"network": "03a98be3-7daa-443f-bb25-4bc6854b396c",
 	"network_cidr": "172.16.0.0/24",
+	"plan": "development",
 	"node_groups": [
 		{
 			"name": "upcloud-go-sdk-unit-test",
@@ -103,6 +104,7 @@ func exampleKubernetesCluster() KubernetesCluster {
 		State: KubernetesClusterStateRunning,
 		UUID:  "0ddab8f4-97c0-4222-91ba-85a4fff7499b",
 		Zone:  "de-fra1",
+		Plan:  "development",
 	}
 }